\name{Cens}
\alias{Cens}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Term to specify the censoring distribution.
}
\description{
Term to specify the censoring distribution.
}
\usage{
Cens(t, ic)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{t}{A vector of observed times.}
  \item{ic}{A vector of cause indicators.}
}
\author{Peng He}