\name{print.crrwt}
\alias{print.crrwt}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Print function for a "crrwt" class object.}
\description{Display major contents of a 'crrwt' class object.}
\usage{\method{print}{crrwt}(x, ...)}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{A 'crrwt' class object obtained from crr.wt function.}
  \item{...}{further arguments passed to or from other methods.}
}
\details{Display parameter estimates as well as their standard errors from a fitted regression model.}
\author{Peng He}
