\name{summary.crrwt}
\alias{summary.crrwt}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Summary function for 'crrwt' class object
}
\description{
Display parameter estimates, standard errors, and results from Wald's tests.
}
\usage{
\method{summary}{crrwt}(object, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{An 'crrwt' class object obtained from crr.wt function.}
  \item{...}{additional arguments affecting the summary produced.}
}
\details{
Results are summarized and structured as a table.
}
\author{
Peng He
}
