\name{wtcrsk-package}
\alias{wtcrsk-package}
\alias{wtcrsk}
\docType{package}
\title{
Competing isks regression models with covariate-adjusted censoring weights
}
\description{
Estimation and regression modeling of subdistribution functions for competing risks data with adjustment for covariate-dependent censoring, as described in He et al. (2015).
}
\details{
\tabular{ll}{
Package: \tab wtcrsk\cr
Type: \tab Package\cr
Version: \tab 1.6\cr
Date: \tab 2015-12-03\cr
License: \tab GPL >= 2\cr
}
}
\author{
Peng He

Maintainer: Peng He <saturnhp@gmail.com>
}

\references{ 
He P, Eriksson F, Scheike TH and Zhang MJ, A proportional hazards regression model for the subdistribution with covariates adjusted censoring weight for competing risks data, Scandinavian Journal of Statistics (2015), Article first published online.
}

\keyword{ package }

