\name{wtcrsk-package}
\alias{wtcrsk-package}
\alias{wtcrsk}
\docType{package}
\title{
Competing risks regression models with covariate-adjusted censoring weight
}
\description{
Estimation and regression modeling of subdistribution functions for competing risks data with adjustment for covariate-dependent censoring, as described in He et al. (2013).
}
\details{
\tabular{ll}{
Package: \tab wtcrsk\cr
Type: \tab Package\cr
Version: \tab 1.2\cr
Date: \tab 2014-04-30\cr
License: \tab GPL >= 2\cr
}
}
\author{
Peng He

Maintainer: Peng He <saturnhp@gmail.com>
}

\references{ 
He P, Scheike TH and Zhang MJ, A proportional hazards regression model for the subdistribution with covariates adjusted censoring weight for competing risks data, Technical Report #61, Division of Biostatistics, Medical College of Wisconsin, November 2013.
}

\keyword{ package }

